modified EndPointHost to use instance
