+++ src/ServiceStack/ServiceStack.csproj
@@ -238,6 +238,7 @@
     <Compile Include="ServiceHost\ServiceRoutes.cs" />
     <Compile Include="WebHost.Endpoints\AppHostBase.cs" />
     <Compile Include="WebHost.Endpoints\AppHostHttpListenerLongRunningBase.cs" />
+    <Compile Include="WebHost.Endpoints\EndpointHostInstance.cs" />
     <Compile Include="WebHost.Endpoints\Extensions\HttpResponseStreamExtensions.cs" />
     <Compile Include="WebHost.Endpoints\Formats\CsvFormat.cs" />
     <Compile Include="WebHost.Endpoints\Extensions\HttpListenerRequestWrapper.cs" />
